refactor(student-addition): extract API business logic
