+++ src/app/api/school/[udise]/students/route.ts
@@ -1,132 +1,32 @@
 import { readFile } from "fs/promises";
-import path from "path";
 
 import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
 
 import { authOptions } from "@/lib/auth";
 import { query } from "@/lib/db";
-import { getDbServiceConfig } from "@/lib/db-service-config";
 import { deriveLmsEnrollmentPeriod } from "@/lib/lms-enrollment-date";
 import {
   ACTIVE_REGISTRATION_MODE,
-  APPROVED_REGISTRATION_MODE,
-  getRegistrationModeHandshake,
-  isRegistrationModeMismatchResponse,
-  PHONE_REGISTRATION_MODE,
-  REGISTRATION_MODE_MISMATCH_MESSAGE,
-  type RegistrationMode,
 } from "@/lib/registration-mode";
 import {
   parseStudentAdditionUpload,
-  type StudentAdditionUploadRowResult,
 } from "@/lib/student-addition-bulk";
 import { requireStudentAdditionAccess } from "@/lib/student-addition-access";
 import {
   validateStudentAdditionInput,
-  type LmsStudentAdditionRow,
-  type StudentAdditionValidationResult,
 } from "@/lib/student-addition-fields";
 
-interface RouteSchool {
-  id: string;
-  code: string;
-  udise_code: string | null;
-  region: string | null;
-  af_school_category: string | null;
-}
-
-interface StudentAdditionAccess {
-  programId: number;
-  actor: {
-    user_id: number | null;
-    email: string;
-    login_type: "google";
-    role: string;
-  };
-}
-
-type DbServiceResult = {
-  row_number: number;
-  status: "created" | "duplicate_in_file" | "already_exists" | "rejected";
-  original?: Record<string, string>;
-};
-
-const EMPTY_TOTALS = {
-  total: 0,
-  created: 0,
-  duplicate_in_file: 0,
-  already_exists: 0,
-  rejected: 0,
-};
-const MAX_STUDENT_ADDITION_UPLOAD_BYTES = 5 * 1024 * 1024;
-const STUDENT_ADDITION_TEMPLATE_FILENAMES: Record<RegistrationMode, string> = {
-  [PHONE_REGISTRATION_MODE]: "NVS_Lakshya_Data_Template_updated_19th_August_2026.xlsx",
-  [APPROVED_REGISTRATION_MODE]: "nvs-student-addition-template.xlsx",
-};
-
-function studentAdditionTemplatePath(mode: RegistrationMode = ACTIVE_REGISTRATION_MODE) {
-  return path.join(
-    process.cwd(),
-    process.env.NODE_ENV === "production" ? ".next/server/assets" : "src/assets",
-    STUDENT_ADDITION_TEMPLATE_FILENAMES[mode],
-  );
-}
-
-function safeFields(value: unknown, keys: string[]) {
-  if (!value || typeof value !== "object" || Array.isArray(value)) return undefined;
-  const record = value as Record<string, unknown>;
-  return Object.fromEntries(keys.filter((key) => key in record).map((key) => [key, record[key]]));
-}
-
-const PHONE_RESTRICTED_RESULT_KEYS = new Set([
-  "pen_number",
-  "g10_roll_no",
-  "annual_family_income",
-  "apaar_id",
-]);
-
-function safePhoneResultFields(value: unknown, mode: RegistrationMode) {
-  if (mode !== PHONE_REGISTRATION_MODE || !value || typeof value !== "object" || Array.isArray(value)) {
-    return value;
-  }
-  return Object.fromEntries(
-    Object.entries(value as Record<string, unknown>)
-      .filter(([key]) => !PHONE_RESTRICTED_RESULT_KEYS.has(key)),
-  );
-}
-
-function safeUpstreamResults(value: unknown, mode: RegistrationMode = ACTIVE_REGISTRATION_MODE) {
-  if (!Array.isArray(value)) return undefined;
-  return value.map((result) => {
-    const safe = safeFields(result, [
-      "row_number", "status", "generated_student_id", "field_errors", "row_errors",
-    ]) ?? {};
-    const record = result as Record<string, unknown>;
-    const duplicateIdentifiers = Array.isArray(record.duplicate_identifiers)
-      ? record.duplicate_identifiers.filter(
-        (identifier): identifier is string =>
-          typeof identifier === "string" &&
-          (mode !== PHONE_REGISTRATION_MODE || !/pen|grade\s*10\s*roll|annual\s*family\s*income/i.test(identifier)),
-      )
-      : undefined;
-    const normalized = safePhoneResultFields(safeFields(record.normalized, [
-      "student_id", "pen_number", "student_name", "g10_roll_no",
-    ]), mode);
-    const existingMatch = safePhoneResultFields(safeFields(record.existing_match, [
-      "matched_identifier", "student_id", "pen_number", "apaar_id", "student_name",
-      "school_name", "school_code", "udise_code", "district", "state", "grade", "program", "stream",
-    ]), mode);
-    const fieldErrors = safePhoneResultFields(safe.field_errors, mode);
-    return {
-      ...safe,
-      ...(fieldErrors ? { field_errors: fieldErrors } : {}),
-      ...(duplicateIdentifiers ? { duplicate_identifiers: duplicateIdentifiers } : {}),
-      ...(normalized ? { normalized } : {}),
-      ...(existingMatch ? { existing_match: existingMatch } : {}),
-    };
-  });
-}
+import {
+  countStudentAdditionTotals,
+  MAX_STUDENT_ADDITION_UPLOAD_BYTES,
+  mergeStudentAdditionResults,
+  proxyStudentAdditionRows,
+  STUDENT_ADDITION_TEMPLATE_FILENAMES,
+  studentAdditionTemplatePath,
+  studentAdditionValidationBody,
+  type StudentAdditionSchool,
+} from "@/lib/student-addition-api";
 
 function isUploadFile(value: FormDataEntryValue | null): value is File {
   return typeof value === "object" &&
@@ -140,31 +40,6 @@ function uploadFilename(file: File): string {
   return file.type.includes("csv") ? "upload.csv" : "upload.xlsx";
 }
 
-function validationResponse(result: StudentAdditionValidationResult) {
-  return NextResponse.json(
-    {
-      error: "Validation failed",
-      totals: { total: 1, created: 0, duplicate_in_file: 0, already_exists: 0, rejected: 1 },
-      results: [
-        {
-          row_number: 1,
-          status: "rejected",
-          generated_student_id: result.generatedStudentId,
-          normalized: {
-            student_name: result.row.student_name ?? "",
-            g10_roll_no: result.row.g10_roll_no ?? "",
-            student_id: result.generatedStudentId,
-          },
-          field_errors: result.fieldErrors,
-          row_errors: result.rowErrors,
-          existing_match: null,
-        },
-      ],
-    },
-    { status: 400 },
-  );
-}
-
 async function resolveSchoolAndAccess(
   session: Parameters<typeof requireStudentAdditionAccess>[0],
   udise: string,
@@ -173,7 +48,7 @@ async function resolveSchoolAndAccess(
     return { response: NextResponse.json({ error: "Unauthorized" }, { status: 401 }) };
   }
 
-  const schools = await query<RouteSchool>(
+  const schools = await query<StudentAdditionSchool>(
     `SELECT
        sch.id,
        sch.code,
@@ -204,83 +79,10 @@ async function resolveRouteContext(params: Promise<{ udise: string }>) {
   return resolveSchoolAndAccess(session, udise);
 }
 
-function countTotals(results: Array<{ status: DbServiceResult["status"] }>) {
-  return results.reduce(
-    (totals, result) => ({
-      ...totals,
-      total: totals.total + 1,
-      [result.status]: totals[result.status] + 1,
-    }),
-    { ...EMPTY_TOTALS },
-  );
-}
-
-// fallow-ignore-next-line complexity
-async function proxyRowsToDbService({
-  access,
-  school,
-  rows,
-  upload,
-  period,
-}: {
-  access: StudentAdditionAccess;
-  school: RouteSchool;
-  rows: LmsStudentAdditionRow[];
-  upload: { id: string; filename: string };
-  period: ReturnType<typeof deriveLmsEnrollmentPeriod>;
-}) {
-  const dbService = getDbServiceConfig();
-  if (!dbService) {
-    return NextResponse.json({ error: "DB Service is not configured" }, { status: 500 });
-  }
-
-  const response = await fetch(`${dbService.baseUrl}/lms/students/bulk-create-with-enrollments`, {
-    method: "POST",
-    headers: dbService.headers,
-    body: JSON.stringify({
-      actor: access.actor,
-      school: { code: school.code, udise_code: school.udise_code },
-      program_id: access.programId,
-      ...getRegistrationModeHandshake(),
-      upload,
-      ...period,
-      rows,
-    }),
-  });
-
-  if (!response.ok) {
-    const upstream = await response.json().catch(() => null) as Record<string, unknown> | null;
-    if (isRegistrationModeMismatchResponse(upstream)) {
-      return NextResponse.json(
-        { error: REGISTRATION_MODE_MISMATCH_MESSAGE },
-        { status: 503 },
-      );
-    }
-    return NextResponse.json(
-      {
-        error: "Student could not be created",
-        ...(upstream?.field_errors ? { field_errors: upstream.field_errors } : {}),
-        ...(upstream?.row_errors ? { row_errors: upstream.row_errors } : {}),
-        ...(upstream?.results ? { results: safeUpstreamResults(upstream.results) } : {}),
-      },
-      { status: response.status },
-    );
-  }
-
-  const body = await response.json() as Record<string, unknown>;
-  return NextResponse.json({
-    ...body,
-    ...(Array.isArray(body.results)
-      ? { results: safeUpstreamResults(body.results) }
-      : {}),
-  }, { status: response.status });
-}
-
-// fallow-ignore-next-line complexity
 async function bulkUploadResponse(
   request: NextRequest,
-  access: StudentAdditionAccess,
-  school: RouteSchool,
+  access: Awaited<ReturnType<typeof requireStudentAdditionAccess>> & { ok: true },
+  school: StudentAdditionSchool,
 ) {
   const form = await request.formData();
   const file = form.get("file");
@@ -315,7 +117,7 @@ async function bulkUploadResponse(
   if (parsed.rows.length === 0) {
     return NextResponse.json(
       {
-        totals: countTotals(parsed.rejectedResults),
+        totals: countStudentAdditionTotals(parsed.rejectedResults),
         results: parsed.rejectedResults,
         ...(parsed.ignoredRows.length > 0 ? { ignored_rows: parsed.ignoredRows } : {}),
       },
@@ -323,7 +125,7 @@ async function bulkUploadResponse(
     );
   }
 
-  const response = await proxyRowsToDbService({
+  const response = await proxyStudentAdditionRows({
     access,
     school,
     rows: parsed.rows,
@@ -333,25 +135,15 @@ async function bulkUploadResponse(
     },
     period,
   });
-  const status = response.status;
-  const body = await response.json();
-  if (!Array.isArray(body.results)) {
-    return NextResponse.json(body, { status });
-  }
-  const dbResults = (body.results ?? []).map((result: DbServiceResult) => ({
-    ...result,
-    original: parsed.originalRows.get(result.row_number) ?? {},
-  }));
-  const results = [...dbResults, ...parsed.rejectedResults].sort(
-    (a, b) => (a.row_number ?? 0) - (b.row_number ?? 0),
-  ) as Array<DbServiceResult | StudentAdditionUploadRowResult>;
-
-  return NextResponse.json({
-    ...body,
-    totals: countTotals(results),
-    results,
-    ...(parsed.ignoredRows.length > 0 ? { ignored_rows: parsed.ignoredRows } : {}),
-  }, { status });
+  return NextResponse.json(
+    mergeStudentAdditionResults({
+      body: response.body,
+      parsedRejectedResults: parsed.rejectedResults,
+      originalRows: parsed.originalRows,
+      ignoredRows: parsed.ignoredRows,
+    }),
+    { status: response.status },
+  );
 }
 
 export async function POST(
@@ -373,8 +165,10 @@ export async function POST(
     rowNumber: 1,
     academicYear: period.academic_year,
   });
-  if (!validation.ok) return validationResponse(validation);
-  return proxyRowsToDbService({
+  if (!validation.ok) {
+    return NextResponse.json(studentAdditionValidationBody(validation), { status: 400 });
+  }
+  const response = await proxyStudentAdditionRows({
     access,
     school,
     rows: [validation.row],
@@ -384,6 +178,7 @@ export async function POST(
     },
     period,
   });
+  return NextResponse.json(response.body, { status: response.status });
 }
 
 export async function GET(
